ui: WorkoutScreen ui changes
