+++ core/ui/designsystem/src/main/res/values/string.xml
@@ -31,7 +31,7 @@
     <string name="settings">Settings</string>
 
     <!-- WorkoutButtons.kt -->
-    <string name="next_workout">Next workout</string>
+    <string name="next_exercise">Next exercise</string>
 
 
     <!--component/topAppBars/ ===================================================================-->
