+++ feature/workout/src/main/java/com/example/feature/workout/workout/component/Buttons.kt
@@ -2,15 +2,16 @@ package com.example.feature.workout.workout.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.core.ui.designsystem.components.button.NextWorkoutButton
+import com.example.core.ui.designsystem.components.button.NextExerciseButton
 import com.example.core.ui.designsystem.components.button.PauseButton
 import com.example.core.ui.designsystem.components.button.StopButton
 import com.example.core.ui.designsystem.components.utils.MySpacerRow
+import com.example.core.utils.itemMaxWidthSmall
 
 @Composable
 internal fun WorkoutButtons(
@@ -19,7 +20,7 @@ internal fun WorkoutButtons(
     onClickNextWorkout: () -> Unit
 ){
     Row(
-        modifier = Modifier.fillMaxWidth(),
+        modifier = Modifier.width(itemMaxWidthSmall),
         horizontalArrangement = Arrangement.Center,
         verticalAlignment = Alignment.CenterVertically
     ) {
@@ -35,7 +36,7 @@ internal fun WorkoutButtons(
 
         MySpacerRow(16.dp)
 
-        NextWorkoutButton(
+        NextExerciseButton(
             onClick = onClickNextWorkout
         )
     }
